refactor(stm32): move the RCC configuration to a dedicated module
